# Backend/docs/swagger.yaml
definitions:
...
  controllers.ErrorMessage:
    properties:
      error:
        example: 'Terjadi kesalahan: data tidak valid'
        type: string
    type: object
...
  controllers.LoginRequest:
    properties:
      password:
        example: rahasia123
        type: string
      username:
        example: admin_utama
        type: string
    type: object
  controllers.LoginSiswaRequest:
    properties:
      device_token:
        example: hp-hafizh-xyz-123
        type: string
      password:
        example: password123
        type: string
      username:
        example: siswa01
        type: string
    type: object
  controllers.LoginSiswaResponse:
    properties:
      data:
        properties:
          role:
            example: siswa
            type: string
          token:
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp...
            type: string
        type: object
      message:
        example: Login berhasil!
        type: string
      status:
        example: success
        type: string
    type: object
...
info:
  contact:
    email: hafizh@example.com
    name: Hafizh
  description: Ini adalah dokumentasi REST API untuk aplikasi absensi dengan fitur
    QR & Geofencing.
  termsOfService: http://swagger.io/terms/
  title: API Sistem Absensi Tempat Les
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Autentikasi untuk Admin dan Guru mengembalikan token JWT
      parameters:
      - description: Kredensial Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil Login
          schema:
            $ref: '#/definitions/controllers.LoginSiswaResponse'
      summary: Login Admin & Guru
      tags:
      - 1. Auth & Keamanan
  /login/siswa:
    post:
      consumes:
      - application/json
      description: Autentikasi siswa dan mendaftarkan ID perangkat (Device Binding)
      parameters:
      - description: Payload Data Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginSiswaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil Login
          schema:
            $ref: '#/definitions/controllers.LoginSiswaResponse'
        "401":
          description: Kredensial Salah
          schema:
            $ref: '#/definitions/controllers.ErrorMessage'
      summary: Login untuk Siswa
      tags:
      - 1. Auth & Keamanan
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"